Refactor: Deduplicate code in graph_slider_ops.c
In most of the operators the code that loops through the FCurve segments and draws the status header were pretty much identical.

To combat that, create abstract helper functions that can be used.

They are not useful in all cases, e.g. the blend to default operator needs a bit of special code.

Pull Request: https://projects.blender.org/blender/blender/pulls/107170